fix(replay): P0 — int32 wrap in ResolveWindowBounds' capture-start chain (33fb586 follow-up)
Worker investigation proved uploaded docs carry a WRAPPED startMs — the true epoch-ms mod 2^32
(owner-proven arithmetically on run sea/dN42ox4Nhd: difference exactly 416 * 2^32). Root cause: the
ResolveWindowBounds input chain narrowed the epoch-scale (~1.7e12, far past int.MaxValue's ~2.147e9)
ServerNowMs-derived value to `int` at its VERY ORIGIN and stored it that way:

- Replay/ReplayCapture.cs: `_combatStartMs` field, the `CombatStartMs` property, `Tick(int nowMs,...)`
  and `SampleAllEntities(int nowMs)` were all `int` — the FIRST stamp
  (`_combatStartMs = nowMs` on the first Tick after Reset()) already wrapped the value permanently
  for the rest of the run. Every later read of `_replay.CombatStartMs` inherited the wrap regardless
  of how correctly downstream math was written.
- Plugin.Replay.cs:212 (now the `nowMs` local in TickReplayCapture): `(int)_services.CombatSnapshot.ServerNowMs`
  — truncated BEFORE being passed into `_replay.Tick(nowMs, dtMs)`.
  Plugin.Replay.cs:434 (now the `upperMs` computation in PrepareReplayDoc):
  `(long)((int)_services.CombatSnapshot.ServerNowMs - _replay.CombatStartMs)` — re-truncated
  ServerNowMs a second, independent time.
  Plugin.Replay.cs:435 (the CapUpper cap arg): `(int)replayUpperCapServerMs - _replay.CombatStartMs`
  — same bug on the boss-phase-cut server timestamp.
  Plugin.Replay.cs:442 (msOffset, now :445): `_replay.CombatStartMs - (int)encounter.StartMs` —
  truncated encounter.StartMs independently before subtracting.
  Plugin.ReplayWindow.cs: `ResolveWindowBounds(int captureStartMs, ...)` — the 33fb586 fix's own
  signature accepted the (already-wrapped) `int` CombatStartMs and added it straight into StartMs/EndMs.

Fix: `long` end to end. `ReplayCapture._combatStartMs`/`CombatStartMs`/`Tick`/`SampleAllEntities`,
`Plugin.Replay.cs`'s `nowMs`/`upperMs`/`msOffset` computations (subtract in `long`, narrow ONLY the
guaranteed-small RESULT for the genuinely-relative `msOffset`), `TickHpTimelines`/
`TickStageBossHpTracks`/`TickEliteHpTracks`'s `nowMs` parameters (Plugin.BossDetection.cs/
Plugin.EliteDetection.cs), and `Plugin.ResolveWindowBounds`'s `captureStartMs` parameter are all
`long` now. `PositionSample.Ms`/`PositionTrackDto.Ms0` stay `int` deliberately — audited and
confirmed they only ever hold genuinely small CAPTURE-RELATIVE deltas (never epoch-scale), matching
their existing, correct design; `PositionUploadDoc.StartMs`/`EndMs` were already `long` and
`PositionJsonWriter`/`PositionCanonicalPayload` already write them via `long`-safe paths (`w.Long`,
`.ToString()`) — no truncation found in the serialization seam itself, only in the plugin-side
inputs feeding it.

Tests: WindowAnchorTests' CaptureStart upgraded from a small int-range constant (1_000_000 — the
exact reason this escaped detection the first time, since it never wraps) to a realistic epoch value
(1_770_000_000_000L); added EpochScaleCaptureStart_NeverWrapsAtInt32Boundary and
OldInt32Truncation_WrapsByAnExactMultipleOf2Pow32_NewFormulaDoesNot (reproduces the OLD `(int)`
truncation via `unchecked((int)...)`, matching the owner's own "off by an exact multiple of 2^32"
proof shape: 412 * 2^32 for this fixture, vs the owner's measured 416 * 2^32). Added
CombatStartMs_StaysEpochScale_NeverWrapsAtInt32Boundary directly on ReplayCapture (the wrap's
ORIGIN) in ReplayCaptureTests.cs. RED evidence: temporarily reverted ONLY ResolveWindowBounds'
signature back to `int captureStartMs` (`sed`, not committed) with the ReplayCapture fix still in
place — the build FAILED with CS1503 "cannot convert from 'long' to 'int'" at the
`ResolveWindowBounds(_replay.CombatStartMs, ...)` call site, concrete proof the type system itself
now rejects the old truncating signature; reverted the sed, confirmed GREEN again before committing.

Full Stellar.CombatMeter.Tests suite: 1135/1135 green (was 1132/1132; +3 new). Release build clean, 0
new warnings. Plugin.Replay.cs grows by +3 lines over its already-flagged 627-line state (630 total,
now 4 over the original 626 baseline) — the minimum needed to correctly document and fix five
distinct narrowing sites in one already-large method; comments kept to one line each throughout. All
other touched files stay well under the 500-LoC guardrail.

Not deployed — coordinator deploys and verifies end-to-end.

+++ Plugin.BossDetection.cs
@@ -441,7 +441,7 @@ public sealed partial class Plugin
     /// alongside the existing sticky <c>killed</c> flag (wire-confirmed death OR the scripted-vanish
     /// inference, both resolved upstream in <see cref="BossStatus"/>). MarkDead is idempotent, so
     /// evaluating both conditions every tick is safe.</summary>
-    private void TickStageBossHpTracks(HpTimelineSampler sampler, long combatStartMs, int nowMs)
+    private void TickStageBossHpTracks(HpTimelineSampler sampler, long combatStartMs, long nowMs)
     {
         if (_stageBosses.Count > 0)
         {
